FAR-84/feat:attraction 기능 구현 test 완료

+++ src/main/java/com/ssafy/faraway/domain/attraction/dto/res/AttractionResponse.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class AttractionResponse {
-    private Long contentId;
+    private int contentId;
     private String title;
     private String addr1;
     private String zipcode;
