+++ src/main/java/com/ssafy/faraway/domain/attraction/repository/SidoRepository.java
@@ -3,5 +3,5 @@ package com.ssafy.faraway.domain.attraction.repository;
 import com.ssafy.faraway.domain.attraction.entity.Sido;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SidoRepository extends JpaRepository<Sido, Long> {
+public interface SidoRepository extends JpaRepository<Sido, Integer> {
 }
